Extracts test setup/teardown to reduce duplication.
